objId is size_t
git-svn-id: file:///home/roberto/Git/svn@7141 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 3532b7cdede157387b247260e1f28e5ffd80f3f2

+++ trunk/xmipp/applications/programs/center_image/center_image_main.cpp
@@ -54,7 +54,7 @@ public:
         std::cout << "LimitShift = " << limitShift << std::endl;
     }
 
-    void processImage(const FileName &fnImg, const FileName &fnImgOut, long int objId)
+    void processImage(const FileName &fnImg, const FileName &fnImgOut, size_t objId)
     {
       Image<double> img;
       img.readApplyGeo(fnImg,mdIn,objId);
